lib/posix-socket: Add UK_LIBC_SYSCALLS guard around send() and recv()
`send()` and `recv()` are defined as wrappers in `posix-socket` to be
used in case they aren't already defined by a libc (such as Musl). Once
a libc is used, there will be a conflict, since the libc would define
the `send()` and `recv()` functions itself.

This commits adds a `#if UK_LIBC_SYSCALLS` guard around the two
functions, such that they will be defined only when a libc is not
present. If a libc (such as Musl) is present, the functions are masked
out.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>